Audio Time-shifter, MP3 Trimmer and concatenator, Background Audio Recorder
This project got it’s motivation from “Heard” iPhone app by Faalsa labs
(http://www.heardapp.com) and decided to create an Android version

The aim of this was a mobile application that will always run
seamlessly in the background and save audio on user-prompt, that could
have been otherwise missed.

Most Audio recorders work on the principle of Record, pause and in some
cases a File editor

The audio will be recorded in the raw PCM format and will be encoded
automatically into mp3 in a separate thread and a separate service in
order not to hog up system memory

The Audio Recording action will be done continuously in a constant loop
of 2 audio files instead of writing to one long file to conserve memory
and maintain the 20MB file size limit. The system utilized the easiest
and least memory-intensive way of recording and encoding audio. The
system was also designed in such a way that it would “run forever”. The
system design is user-friendly and will also be upgradable due to the
flexibility of Android OS.
Reusable Components:
•	Audio Recorder
•	MP3 Audio Concatenation
•	MP3 Trimmer
•	Multi-threading and Background activities
•	Encoder from PCM to MP3

NOTE: As no process can run on android indefinitely, but as long as the
phone is active, the system will work as designed
Pardon the fact there are fewer classes, enjoy, share and please give
credit to Uraroji for the MP3 LAME Encoder